Updating the custom reticle design pattern to allow swappable reticles on the poke interactor and other future interactors as well
