+++ com.microsoft.mrtk.input/Interactors/InteractorVisuals/MRTKBaseReticleVisual.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94d0a3cfa70939b4c802013c695e5d72
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
